fix(settings) - Relocate and correct database settings module
